requirements.txt
gradio>=1.6.4
Keras>=2.4.3
tensorflow>=2.4.1
matplotlib>=3.3.3
nltk>=3.5
numpy>=1.19.5
pandas>=0.19
regex>=2020.6.8
scikit-learn>=0.24.1
seaborn>=0.10.1
spacy>=3.0.5
PyYAML
fastapi
uvicorn
pydantic
fasttext
